Implemented species edit command
(Species edit command does not work for images yet)
I'm really not proud of this. The species edit command has shown me holes in the design of EditableDocument, and my solution for preserving _id as a hidden trait in documents is hacky at best. I plan on completely redesigning EditableDocument eventually.